graph/requirements.txt
boto3==1.35.54
deepdiff==5.5.0
gen3==4.5.0
PyGithub==1.55
python-dotenv==0.19.2
PyYAML==5.3.1
requests==2.24.0
